+++ OpinionatedLeagueFan.py
@@ -1,61 +1,36 @@
-from dotenv import load_dotenv
-
 import os
-import random
 import time
 import datetime
+import random
 import pickle
-import openai
-import praw
 
-load_dotenv()
+import vardata
+import InboxReply
 
-# dev mode will skip posting and only output to console
-dev_mode = 0
-
-# set up OpenAI API
-openai.api_key = os.getenv('OPENAI_API_KEY')
-model_temperature=1.3
+# Set the prompt for public replies
 prompt = "You are an opinionated Rugby League fan on Reddit. \
-        The " + datetime.date.today().strftime("%Y") + " season has started and you don't follow any one particular team - you love all footy. Most others are die hard fans of one particular team, however. \
-        You have poor grammar and spelling, and you like to use shorthand often. \
-        You speak in Australian English. \
-        You like to argue with other users and are not willing to accept that you might be wrong. \
-        Assume that other users are already well-informed of the rules and systems in place in Rugby League. \
-        Some posts that you respond to may not be specifically about Rugby League. \
-        Do not respond to comments as if they were directed at you personally. If comments are directed at an individual, do not respond as if it were directed at you. \
-        Do not try to moderate discussions. \
-        If there is not enough context to the post, simply agree."
+        The " + datetime.date.today().strftime("%Y") + " season has started and you don't follow any one particular team - you love all footy. Most others are die hard fans of one particular team. \
+        Do not respond directly to comments as they're rarely directed at you. \
+        You speak in Australian English."
 
-# set up Reddit API credentials
-reddit = praw.Reddit(client_id=os.getenv('REDDIT_CLIENT_ID'),
-                     client_secret=os.getenv('REDDIT_CLIENT_SECRET'),
-                     user_agent='Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36',
-                     username=os.getenv('REDDIT_USERNAME'),
-                     password=os.getenv('REDDIT_PASSWORD'))
-# set the subreddit to monitor
-subreddit = reddit.subreddit('nrl')
-self = reddit.redditor(os.getenv('REDDIT_USERNAME'))
-
-# keep track of the last post ID processed
-picklefile = 'post-tracking.pk'
-
-with open(picklefile) as f:
-    try:
-        last_post_id, last_comment_id = pickle.load(f)
-    except Exception:
-        # touch the picklefile if it doesn't exist
-        open(picklefile, 'a').close()
-        # now load empty values to work with
-        last_post_id = None
-        last_comment_id = None
+if os.path.exists(vardata.picklefile):
+    last_post_id, last_comment_id = pickle.load(vardata.picklefile)
+else:
+    # touch the picklefile if it doesn't exist
+    open(vardata.picklefile, 'w+').close()
+    # now load empty values to work with
+    last_post_id = None
+    last_comment_id = None
 
 # monitor for new posts
 while True:
     # Only run during common hours
-    if datetime.datetime.now().hour < 6 | datetime.datetime.now().hour > 23:
+    if vardata.dev_mode == 0 & datetime.datetime.now().hour < 6 | datetime.datetime.now().hour > 23:
         print('Outside of operating hours')
         break
+
+    # Handle any inbox replies before making new ones
+    InboxReply.checkInboxReplies()
 
     # randomly select what we're going to reply to
     # 0: comments
@@ -66,39 +41,45 @@ while True:
         match select:
             case 0:
                 # get the latest comment
-                new_comments = subreddit.comments(limit=1)
+                new_comments = vardata.subreddit.comments(limit=1)
 
                 for comment in new_comments:
-                    if comment.author.id == self.id:
+                    if comment.author.id == vardata.self.id:
                         # Skip our own comments
                         raise Exception('Comment is my own, skipping.')
                     if len(comment.body.split()) < 5:
                         # too little words to be a useful comment
                         raise Exception('Comment is not long enough, skipping.')
+                    if comment.author.id in vardata.blacklist:
+                        # Skip Blacklisted user comments
+                        raise Exception('Comment author is blacklisted, skipping.')
 
                     last_comment_id = comment.id
                     print('New comment: ' + comment.body)
 
-                    response = openai.ChatCompletion.create(
+                    response = vardata.openai.ChatCompletion.create(
                         model="gpt-3.5-turbo",
-                        temperature=model_temperature,
+                        temperature=vardata.model_temperature,
+                        # max_tokens=random.randint(5,100),
+                        frequency_penalty=vardata.model_frequency_penalty,
                         messages=[
                                 {"role": "system", "content": prompt},
-                                {"role": "user", "content": "Respond to the following comment: " + comment.body},
+                                {"role": "user", "content": "The thread title is: \"" + comment.submission.title + "\". Respond to the following comment: \"" + comment.body + "\""},
                             ]
                         )
 
                     # wait a random interval of at least 2 minutes to a maximum of 7 before posting
                     print('Replying with: \n' + response['choices'][0]['message']['content'] + '\n')
-                    if dev_mode == 0:
-                        print('Waiting to post...')
-                        time.sleep(int(2) + random.randint(0,300))
+                    if vardata.dev_mode == 0:
+                        waitBeforePost = int(120) + random.randint(0,300)
+                        print('Waiting ' + str(waitBeforePost) + 's to post...')
+                        time.sleep(waitBeforePost)
                         comment.reply(response['choices'][0]['message']['content'])
                         print('Posted\n')
 
             case 1:
                 # get the latest post
-                new_posts = subreddit.new(limit=1)
+                new_posts = vardata.subreddit.new(limit=1)
 
                 # process each new post
                 for post in new_posts:
@@ -106,10 +87,10 @@ while True:
                     if last_post_id and post.id == last_post_id:
                         # skip posts that have already been processed
                         raise Exception('Post is not new, skipping.')
-                    if post.author.id == "AutoModerator":
-                        # Skip AutoMod posts
-                        raise Exception('Post is by AutoMod, skipping.')
-                    if post.author.id == self.id:
+                    if post.author.id in vardata.blacklist:
+                        # Skip Blacklisted user posts
+                        raise Exception('Post author is blacklisted, skipping.')
+                    if post.author.id == vardata.self.id:
                         # Skip our own posts
                         raise Exception('Post is my own, skipping.')
                     if post.title == 'Fantasy Football and SuperCoach Thread':
@@ -127,20 +108,23 @@ while True:
                     print(f'New post: {post.title}')
                     print(f'Body: {post.selftext}')
 
-                    response = openai.ChatCompletion.create(
+                    response = vardata.openai.ChatCompletion.create(
                         model="gpt-3.5-turbo",
-                        temperature=model_temperature,
+                        temperature=vardata.model_temperature,
+                        max_tokens=random.randint(5,100),
+                        frequency_penalty=vardata.model_frequency_penalty,
                         messages=[
                                 {"role": "system", "content": prompt},
-                                {"role": "user", "content": "Respond to the following new post: " + post.selftext},
+                                {"role": "user", "content": "The thread title is: \"" + post.title + "\". Respond to the following new post: " + post.selftext + "\""},
                             ]
                         )
                     
                     # wait a random interval of at least 2 minutes to a maximum of 7 before posting
                     print('Repling with: \n' + response['choices'][0]['message']['content'] + '\n')
-                    if dev_mode == 0:
-                        print('Waiting to post...')
-                        time.sleep(int(2) + random.randint(0,300))
+                    if vardata.dev_mode == 0:
+                        waitBeforePost = int(120) + random.randint(0,300)
+                        print('Waiting ' + str(waitBeforePost) + 's to post...')
+                        time.sleep(waitBeforePost)
                         post.reply(response['choices'][0]['message']['content'])
                         print('Posted\n')
 
@@ -148,7 +132,7 @@ while True:
                     last_post_id = post.id
         
         # save the current state
-        with open(picklefile, 'wb') as f:
+        with open(vardata.picklefile, 'wb') as f:
             pickle.dump([last_post_id, last_comment_id], f)
 
         # wait 60 seconds before checking for new posts again