+++ changelog.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-03-16
+### Added
+- Inbox replies
+- User blacklist
+### Changed
+- Prompting for posts and comments
+- Refactored to DRY format for re-using variables
+### Fixed
+- Picklefile creation when the file does not exist
+- Typo for random reply intervals (minutes instead of seconds)
 ## [1.0.1] - 2023-03-15
 ### Added
 - Dev mode